backend/requirements.txt
# Django Core (compatible with Python 3.9)
django==4.2.15
djangorestframework==3.15.2
django-cors-headers==4.3.1

# LangChain for RAG
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.17

# Vector Store and Embeddings
faiss-cpu==1.8.0
sentence-transformers==3.0.1

# Document Processing
pypdf==4.2.0
python-docx==1.1.2
unstructured==0.14.10
python-magic==0.4.27

# Environment and Configuration
python-dotenv==1.0.1
Pillow==10.4.0

# Background Tasks (for async processing)
celery==5.3.4
redis==5.0.7

# Development and Utilities
django-extensions==3.2.3
ipython==8.18.1
